# Improvements in Wazuh dahboard 4.6.0-2.8.0 package