--- a/stack/dashboard/deb/debian/postinst
+++ b/stack/dashboard/deb/debian/postinst
@@ -36,10 +36,14 @@ case "$1" in
                 service wazuh-dashboard restart > /dev/null 2>&1
             fi
         fi
-        if [ ! -f "${CONFIG_DIR}"/opensearch_dashboards.keystore ]; then
+        # Move keystore file if upgrade (file exists in install dir in <= 4.6.0)
+        if [ -f "${INSTALLATION_DIR}"/opensearch_dashboards.keystore ]; then
+            mv "${INSTALLATION_DIR}"/opensearch_dashboards.keystore "${CONFIG_DIR}"/opensearch_dashboards.keystore
+        elif [ ! -f "${CONFIG_DIR}"/opensearch_dashboards.keystore ]; then
             runuser "${NAME}" --shell="/bin/bash" --command="${INSTALLATION_DIR}/bin/opensearch-dashboards-keystore create" > /dev/null 2>&1
             runuser "${NAME}" --shell="/bin/bash" --command="echo kibanaserver | ${INSTALLATION_DIR}/bin/opensearch-dashboards-keystore add opensearch.username --stdin" > /dev/null 2>&1
             runuser "${NAME}" --shell="/bin/bash" --command="echo kibanaserver | ${INSTALLATION_DIR}/bin/opensearch-dashboards-keystore add opensearch.password --stdin" > /dev/null 2>&1
+            chmod 640 "${CONFIG_DIR}"/opensearch_dashboards.keystore
         fi
     ;;