--- a/stack/dashboard/deb/debian/postinst
+++ b/stack/dashboard/deb/debian/postinst
@@ -37,8 +37,8 @@ case "$1" in
             fi
         fi
         # Move keystore file if upgrade (file exists in install dir in <= 4.6.0)
-        if [ -f "${INSTALLATION_DIR}"/opensearch_dashboards.keystore ]; then
-            mv "${INSTALLATION_DIR}"/opensearch_dashboards.keystore "${CONFIG_DIR}"/opensearch_dashboards.keystore
+        if [ -f "${INSTALLATION_DIR}"/config/opensearch_dashboards.keystore ]; then
+            mv "${INSTALLATION_DIR}"/config/opensearch_dashboards.keystore "${CONFIG_DIR}"/opensearch_dashboards.keystore
         elif [ ! -f "${CONFIG_DIR}"/opensearch_dashboards.keystore ]; then
             runuser "${NAME}" --shell="/bin/bash" --command="${INSTALLATION_DIR}/bin/opensearch-dashboards-keystore create" > /dev/null 2>&1
             runuser "${NAME}" --shell="/bin/bash" --command="echo kibanaserver | ${INSTALLATION_DIR}/bin/opensearch-dashboards-keystore add opensearch.username --stdin" > /dev/null 2>&1