|Related issue|
|---|
|Related https://github.com/wazuh/wazuh-packages/issues/2420#top|


## Description

This PR adds the necessary changes so that the Wazuh dashboard 4.6.0-2.8.0 package can use IPs in the `opensearch.hosts` configuration, and also moves the OpenSearch keystore to the Wazuh dashboard configuration directory

## Logs example

- https://ci.wazuh.info/view/Packages/job/Packages_builder_tier/3142/
- https://ci.wazuh.info/job/Packages_builder/164381/
- https://ci.wazuh.info/job/Packages_builder/164412/
- https://ci.wazuh.info/job/Packages_builder/164413/
- https://ci.wazuh.info/job/Test_stack_tier/83/

<details><summary>Upgrade in Debian 11 from 4.5.2 to 4.6.0</summary>

```
root@debian11:/home/vagrant# curl -sO https://packages.wazuh.com/4.5/wazuh-install.sh && sudo bash ./wazuh-install.sh -a -i
06/09/2023 17:44:19 INFO: Starting Wazuh installation assistant. Wazuh version: 4.5.2
06/09/2023 17:44:19 INFO: Verbose logging redirected to /var/log/wazuh-install.log
06/09/2023 17:44:24 WARNING: Hardware and system checks ignored.
06/09/2023 17:44:24 INFO: Wazuh web interface port will be 443.
06/09/2023 17:44:25 INFO: --- Dependencies ----
06/09/2023 17:44:25 INFO: Installing apt-transport-https.
06/09/2023 17:44:26 INFO: Installing software-properties-common.
06/09/2023 17:44:31 INFO: Wazuh repository added.
06/09/2023 17:44:31 INFO: --- Configuration files ---
06/09/2023 17:44:31 INFO: Generating configuration files.
06/09/2023 17:44:31 INFO: Created wazuh-install-files.tar. It contains the Wazuh cluster key, certificates, and passwords necessary for installation.
06/09/2023 17:44:31 INFO: --- Wazuh indexer ---
06/09/2023 17:44:31 INFO: Starting Wazuh indexer installation.
06/09/2023 17:45:11 INFO: Wazuh indexer installation finished.
06/09/2023 17:45:11 INFO: Wazuh indexer post-install configuration finished.
06/09/2023 17:45:11 INFO: Starting service wazuh-indexer.
06/09/2023 17:45:24 INFO: wazuh-indexer service started.
06/09/2023 17:45:24 INFO: Initializing Wazuh indexer cluster security settings.
06/09/2023 17:45:34 INFO: Wazuh indexer cluster initialized.
06/09/2023 17:45:34 INFO: --- Wazuh server ---
06/09/2023 17:45:34 INFO: Starting the Wazuh manager installation.
06/09/2023 17:46:00 INFO: Wazuh manager installation finished.
06/09/2023 17:46:00 INFO: Starting service wazuh-manager.
06/09/2023 17:46:14 INFO: wazuh-manager service started.
06/09/2023 17:46:14 INFO: Starting Filebeat installation.
06/09/2023 17:46:17 INFO: Filebeat installation finished.
06/09/2023 17:46:18 INFO: Filebeat post-install configuration finished.
06/09/2023 17:46:18 INFO: Starting service filebeat.
06/09/2023 17:46:18 INFO: filebeat service started.
06/09/2023 17:46:18 INFO: --- Wazuh dashboard ---
06/09/2023 17:46:18 INFO: Starting Wazuh dashboard installation.
06/09/2023 17:46:43 INFO: Wazuh dashboard installation finished.
06/09/2023 17:46:43 INFO: Wazuh dashboard post-install configuration finished.
06/09/2023 17:46:43 INFO: Starting service wazuh-dashboard.
06/09/2023 17:46:43 INFO: wazuh-dashboard service started.
06/09/2023 17:47:00 INFO: Initializing Wazuh dashboard web application.
06/09/2023 17:47:01 INFO: Wazuh dashboard web application initialized.
06/09/2023 17:47:01 INFO: --- Summary ---
06/09/2023 17:47:01 INFO: You can access the web interface https://<wazuh-dashboard-ip>:443
    User: admin
    Password: H.69t?XfoK.W2QS9y4pxU1t1P+DS7v7Q
06/09/2023 17:47:01 INFO: Installation finished.
root@debian11:/home/vagrant# systemctl stop filebeat
systemctl stop wazuh-dashboard
root@debian11:/home/vagrant# curl -X PUT "https://localhost:9200/_cluster/settings"  -u admin:H.69t?XfoK.W2QS9y4pxU1t1P+DS7v7Q -k -H 'Content-Type: application/json' -d'
{
  "persistent": {
    "cluster.routing.allocation.enable": "primaries"
  }
}
'
{"acknowledged":true,"persistent":{"cluster":{"routing":{"allocation":{"enable":"primaries"}}}},"transient":{}}root@debian11:/home/vagrant# 
root@debian11:/home/vagrant# curl -X POST "https://localhost:9200/_flush/synced" -u admin:H.69t?XfoK.W2QS9y4pxU1t1P+DS7v7Q -k
{"_shards":{"total":7,"successful":7,"failed":0}}root@debian11:/home/vagrant# 
root@debian11:/home/vagrant# systemctl stop wazuh-indexer
root@debian11:/home/vagrant# apt install ./wazuh-indexer_4.6.0-wp.2420_amd64.deb 
Reading package lists... Done
Building dependency tree... Done
Reading state information... Done
Note, selecting 'wazuh-indexer' instead of './wazuh-indexer_4.6.0-wp.2420_amd64.deb'
The following packages will be upgraded:
  wazuh-indexer
1 upgraded, 0 newly installed, 0 to remove and 56 not upgraded.
Need to get 0 B/684 MB of archives.
After this operation, 1,739 kB disk space will be freed.
Get:1 /home/vagrant/wazuh-indexer_4.6.0-wp.2420_amd64.deb wazuh-indexer amd64 4.6.0-1 [684 MB]
Reading changelogs... Done
(Reading database ... 186868 files and directories currently installed.)
Progress: [ 20%] [###########################################............................................................................................................................................................................] 
Setting up wazuh-indexer (4.6.0-1) ...#######################............................................................................................................................................................................] 
Installing new version of config file /etc/wazuh-indexer/opensearch-notifications-core/notifications-core.yml ...
Installing new version of config file /etc/wazuh-indexer/opensearch-performance-analyzer/log4j2.xml ...
Installing new version of config file /etc/wazuh-indexer/opensearch-performance-analyzer/opensearch_security.policy ...
Installing new version of config file /etc/wazuh-indexer/opensearch-performance-analyzer/rca.conf ...
Installing new version of config file /etc/wazuh-indexer/opensearch-performance-analyzer/rca_cluster_manager.conf ...
Installing new version of config file /etc/wazuh-indexer/opensearch-performance-analyzer/rca_idle_cluster_manager.conf ...
Installing new version of config file /etc/wazuh-indexer/opensearch-security/config.yml ...
root@debian11:/home/vagrant# systemctl daemon-reload
systemctl enable wazuh-indexer
systemctl start wazuh-indexer
root@debian11:/home/vagrant# curl -k -u admin:H.69t?XfoK.W2QS9y4pxU1t1P+DS7v7Q https://localhost:9200/_cat/nodes?v
ip        heap.percent ram.percent cpu load_1m load_5m load_15m node.role node.roles                                        cluster_manager name
127.0.0.1           12          97  10    0.57    0.46     0.25 dimr      cluster_manager,data,ingest,remote_cluster_client *               node-1
root@debian11:/home/vagrant# curl -X PUT "https://localhost:9200/_cluster/settings" -u admin:H.69t?XfoK.W2QS9y4pxU1t1P+DS7v7Q -k -H 'Content-Type: application/json' -d'
{
  "persistent": {
    "cluster.routing.allocation.enable": "all"
  }
}
'
{"acknowledged":true,"persistent":{"cluster":{"routing":{"allocation":{"enable":"all"}}}},"transient":{}}root@debian11:/home/vagrant# 
root@debian11:/home/vagrant# curl -k -u admin:H.69t?XfoK.W2QS9y4pxU1t1P+DS7v7Q https://localhost:9200/_cat/nodes?v
ip        heap.percent ram.percent cpu load_1m load_5m load_15m node.role node.roles                                        cluster_manager name
127.0.0.1           15          97   1    0.27    0.39     0.23 dimr      cluster_manager,data,ingest,remote_cluster_client *               node-1
root@debian11:/home/vagrant# 
root@debian11:/home/vagrant# apt install ./wazuh-manager_4.6.0-wp.2420_amd64.deb 
Reading package lists... Done
Building dependency tree... Done
Reading state information... Done
Note, selecting 'wazuh-manager' instead of './wazuh-manager_4.6.0-wp.2420_amd64.deb'
Suggested packages:
  expect
The following packages will be upgraded:
  wazuh-manager
1 upgraded, 0 newly installed, 0 to remove and 56 not upgraded.
Need to get 0 B/171 MB of archives.
After this operation, 1,704 kB of additional disk space will be used.
Get:1 /home/vagrant/wazuh-manager_4.6.0-wp.2420_amd64.deb wazuh-manager amd64 4.6.0-1 [171 MB]
Reading changelogs... Done        
(Reading database ... 186872 files and directories currently installed.)
Preparing to unpack .../wazuh-manager_4.6.0-wp.2420_amd64.deb ...
Unpacking wazuh-manager (4.6.0-1) over (4.5.2-1) ...
Setting up wazuh-manager (4.6.0-1) ...
root@debian11:/home/vagrant# curl -s https://packages.wazuh.com/4.x/filebeat/wazuh-filebeat-0.2.tar.gz | sudo tar -xvz -C /usr/share/filebeat/module
sudo: unable to resolve host debian11: Name or service not known
wazuh/alerts/
wazuh/alerts/config/
wazuh/alerts/config/alerts.yml
wazuh/alerts/manifest.yml
wazuh/alerts/ingest/
wazuh/alerts/ingest/pipeline.json
wazuh/archives/
wazuh/archives/config/
wazuh/archives/config/archives.yml
wazuh/archives/manifest.yml
wazuh/archives/ingest/
wazuh/archives/ingest/pipeline.json
wazuh/module.yml
root@debian11:/home/vagrant# curl -so /etc/filebeat/wazuh-template.json https://raw.githubusercontent.com/wazuh/wazuh/4.6.0/extensions/elasticsearch/7.x/wazuh-template.json
root@debian11:/home/vagrant# chmod go+r /etc/filebeat/wazuh-template.json
root@debian11:/home/vagrant# systemctl daemon-reload
systemctl enable filebeat
systemctl start filebeat
Synchronizing state of filebeat.service with SysV service script with /lib/systemd/systemd-sysv-install.
Executing: /lib/systemd/systemd-sysv-install enable filebeat
root@debian11:/home/vagrant# apt install ./wazuh-dashboard_4.6.0-1_amd64.deb ^C
root@debian11:/home/vagrant# ls -l /usr/share/wazuh-dashboard/config/opensearch_dashboards.keystore 
-rw-r--r-- 1 wazuh-dashboard wazuh-dashboard 254 Sep  6 17:46 /usr/share/wazuh-dashboard/config/opensearch_dashboards.keystore
You have mail in /var/mail/root
root@debian11:/home/vagrant# sha1sum /usr/share/wazuh-dashboard/config/opensearch_dashboards.keystore 
91fb111839fb819f29e77dcdffee07eef1582f59  /usr/share/wazuh-dashboard/config/opensearch_dashboards.keystore
root@debian11:/home/vagrant# apt install ./wazuh-dashboard_4.6.0-1_amd64.deb 
Reading package lists... Done
Building dependency tree... Done
Reading state information... Done
Note, selecting 'wazuh-dashboard' instead of './wazuh-dashboard_4.6.0-1_amd64.deb'
The following packages will be upgraded:
  wazuh-dashboard
1 upgraded, 0 newly installed, 0 to remove and 56 not upgraded.
Need to get 0 B/179 MB of archives.
After this operation, 152 MB of additional disk space will be used.
Get:1 /home/vagrant/wazuh-dashboard_4.6.0-1_amd64.deb wazuh-dashboard amd64 4.6.0-1 [179 MB]
Reading changelogs... Done
(Reading database ... 186885 files and directories currently installed.)
Preparing to unpack .../wazuh-dashboard_4.6.0-1_amd64.deb ...
Unpacking wazuh-dashboard (4.6.0-1) over (4.5.2-1) ...
Setting up wazuh-dashboard (4.6.0-1) ...
Installing new version of config file /etc/systemd/system/wazuh-dashboard.service ...

Configuration file '/etc/wazuh-dashboard/opensearch_dashboards.yml'
 ==> Modified (by you or by a script) since installation.
 ==> Package distributor has shipped an updated version.
   What would you like to do about it ?  Your options are:
    Y or I  : install the package maintainer's version
    N or O  : keep your currently-installed version
      D     : show the differences between the versions
      Z     : start a shell to examine the situation
 The default action is to keep your current version.
*** opensearch_dashboards.yml (Y/I/N/O/D/Z) [default=N] ? N
root@debian11:/home/vagrant# ls -l /usr/share/wazuh-dashboard/config/opensearch_dashboards.keystore 
ls: cannot access '/usr/share/wazuh-dashboard/config/opensearch_dashboards.keystore': No such file or directory
root@debian11:/home/vagrant# ls -l /etc/wazuh-dashboard/opensearch_dashboards.keystore 
-rw-r--r-- 1 wazuh-dashboard wazuh-dashboard 254 Sep  6 17:46 /etc/wazuh-dashboard/opensearch_dashboards.keystore
root@debian11:/home/vagrant# sha1sum /etc/wazuh-dashboard/opensearch_dashboards.keystore 
91fb111839fb819f29e77dcdffee07eef1582f59  /etc/wazuh-dashboard/opensearch_dashboards.keystore
oot@debian11:/home/vagrant# cat /etc/wazuh-dashboard/opensearch_dashboards.yml
server.host: 0.0.0.0
opensearch.hosts: https://127.0.0.1:9200
server.port: 443
opensearch.ssl.verificationMode: certificate
# opensearch.username: kibanaserver
# opensearch.password: kibanaserver
opensearch.requestHeadersAllowlist: ["securitytenant","Authorization"]
opensearch_security.multitenancy.enabled: false
opensearch_security.readonly_mode.roles: ["kibana_read_only"]
server.ssl.enabled: true
server.ssl.key: "/etc/wazuh-dashboard/certs/wazuh-dashboard-key.pem"
server.ssl.certificate: "/etc/wazuh-dashboard/certs/wazuh-dashboard.pem"
opensearch.ssl.certificateAuthorities: ["/etc/wazuh-dashboard/certs/root-ca.pem"]
uiSettings.overrides.defaultRoute: /app/wazuh
opensearch_security.cookie.secure: true
root@debian11:/home/vagrant# ls -l /etc/wazuh-dashboard/certs/
total 12
-r-------- 1 wazuh-dashboard wazuh-dashboard 1204 Sep  6 17:44 root-ca.pem
-r-------- 1 wazuh-dashboard wazuh-dashboard 1704 Sep  6 17:44 wazuh-dashboard-key.pem
-r-------- 1 wazuh-dashboard wazuh-dashboard 1245 Sep  6 17:44 wazuh-dashboard.pem
root@debian11:/home/vagrant# systemctl daemon-reload
systemctl enable wazuh-dashboard
systemctl start wazuh-dashboard
root@debian11:/home/vagrant# systemctl status wazuh-dashboard
● wazuh-dashboard.service - wazuh-dashboard
     Loaded: loaded (/etc/systemd/system/wazuh-dashboard.service; enabled; vendor preset: enabled)
     Active: active (running) since Wed 2023-09-06 17:56:17 UTC; 1min 47s ago
   Main PID: 100566 (node)
      Tasks: 11 (limit: 4675)
     Memory: 205.4M
        CPU: 3.895s
     CGroup: /system.slice/wazuh-dashboard.service
             └─100566 /usr/share/wazuh-dashboard/node/bin/node --no-warnings --max-http-header-size=65536 --unhandled-rejections=warn /usr/share/wazuh-dashboard/src/cli/dist -c /etc/wazuh-dashboard/opensearch_dashboards.yml

Sep 06 17:56:21 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:21Z","tags":["info","savedobjects-service"],"pid":100566,"message":"Waiting until all OpenSearch nodes are compatible with OpenSearch D>
Sep 06 17:56:21 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:21Z","tags":["info","savedobjects-service"],"pid":100566,"message":"Starting saved objects migrations"}
Sep 06 17:56:21 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:21Z","tags":["info","savedobjects-service"],"pid":100566,"message":"Detected mapping change in \"properties.visualization-visbuilder\""}
Sep 06 17:56:21 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:21Z","tags":["info","savedobjects-service"],"pid":100566,"message":"Creating index .kibana_2."}
Sep 06 17:56:21 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:21Z","tags":["info","savedobjects-service"],"pid":100566,"message":"Migrating .kibana_1 saved objects to .kibana_2"}
Sep 06 17:56:21 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:21Z","tags":["info","savedobjects-service"],"pid":100566,"message":"Pointing alias .kibana to .kibana_2."}
Sep 06 17:56:21 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:21Z","tags":["info","savedobjects-service"],"pid":100566,"message":"Finished in 224ms."}
Sep 06 17:56:21 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:21Z","tags":["info","plugins-system"],"pid":100566,"message":"Starting [44] plugins: [alertingDashboards,usageCollection,opensearchDash>
Sep 06 17:56:22 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:22Z","tags":["listening","info"],"pid":100566,"message":"Server running at https://0.0.0.0:443"}
Sep 06 17:56:22 debian11 opensearch-dashboards[100566]: {"type":"log","@timestamp":"2023-09-06T17:56:22Z","tags":["info","http","server","OpenSearchDashboards"],"pid":100566,"message":"http server running at https://0.0.0.0:443"}
```

</details>

<details><summary>Upgrade in CentOS 7 from 4.5.2 to 4.6.0</summary>

```
[root@centos7 vagrant]# curl -sO https://packages.wazuh.com/4.5/wazuh-install.sh && sudo bash ./wazuh-install.sh -a
06/09/2023 16:04:05 INFO: Starting Wazuh installation assistant. Wazuh version: 4.5.2
06/09/2023 16:04:05 INFO: Verbose logging redirected to /var/log/wazuh-install.log
06/09/2023 16:04:09 INFO: --- Dependencies ---
06/09/2023 16:04:09 INFO: Installing lsof.
06/09/2023 16:04:09 INFO: Wazuh web interface port will be 443.
06/09/2023 16:04:10 INFO: Wazuh repository added.
06/09/2023 16:04:10 INFO: --- Configuration files ---
06/09/2023 16:04:10 INFO: Generating configuration files.
06/09/2023 16:04:10 INFO: Created wazuh-install-files.tar. It contains the Wazuh cluster key, certificates, and passwords necessary for installation.
06/09/2023 16:04:10 INFO: --- Wazuh indexer ---
06/09/2023 16:04:10 INFO: Starting Wazuh indexer installation.
06/09/2023 16:05:28 INFO: Wazuh indexer installation finished.
06/09/2023 16:05:28 INFO: Wazuh indexer post-install configuration finished.
06/09/2023 16:05:28 INFO: Starting service wazuh-indexer.
06/09/2023 16:05:35 INFO: wazuh-indexer service started.
06/09/2023 16:05:35 INFO: Initializing Wazuh indexer cluster security settings.
06/09/2023 16:05:45 INFO: Wazuh indexer cluster initialized.
06/09/2023 16:05:45 INFO: --- Wazuh server ---
06/09/2023 16:05:45 INFO: Starting the Wazuh manager installation.
06/09/2023 16:06:15 INFO: Wazuh manager installation finished.
06/09/2023 16:06:15 INFO: Starting service wazuh-manager.
06/09/2023 16:06:28 INFO: wazuh-manager service started.
06/09/2023 16:06:28 INFO: Starting Filebeat installation.
06/09/2023 16:06:33 INFO: Filebeat installation finished.
06/09/2023 16:06:34 INFO: Filebeat post-install configuration finished.
06/09/2023 16:06:34 INFO: Starting service filebeat.
06/09/2023 16:06:34 INFO: filebeat service started.
06/09/2023 16:06:34 INFO: --- Wazuh dashboard ---
06/09/2023 16:06:34 INFO: Starting Wazuh dashboard installation.
06/09/2023 16:07:25 INFO: Wazuh dashboard installation finished.
06/09/2023 16:07:25 INFO: Wazuh dashboard post-install configuration finished.
06/09/2023 16:07:25 INFO: Starting service wazuh-dashboard.
06/09/2023 16:07:25 INFO: wazuh-dashboard service started.
06/09/2023 16:07:39 INFO: Initializing Wazuh dashboard web application.
06/09/2023 16:07:41 INFO: Wazuh dashboard web application initialized.
06/09/2023 16:07:41 INFO: --- Summary ---
06/09/2023 16:07:41 INFO: You can access the web interface https://<wazuh-dashboard-ip>:443
    User: admin
    Password: SWHH*sEk5qZB0weCwNTp0oFhvAT?ew2E
06/09/2023 16:07:41 INFO: Installation finished.
[root@centos7 vagrant]# ls -l /usr/share/wazuh-dashboard/config/opensearch_dashboards.keystore 
-rw-r--r--. 1 wazuh-dashboard wazuh-dashboard 254 Sep  6 16:07 /usr/share/wazuh-dashboard/config/opensearch_dashboards.keystore
[root@centos7 vagrant]# wget https://packages-dev.wazuh.com/warehouse/test/4.6/rpm/wazuh-indexer-4.6.0-wp.2420.x86_64.rpm
--2023-09-06 16:22:23--  https://packages-dev.wazuh.com/warehouse/test/4.6/rpm/wazuh-indexer-4.6.0-wp.2420.x86_64.rpm
Resolving packages-dev.wazuh.com (packages-dev.wazuh.com)... 52.84.66.65, 52.84.66.16, 52.84.66.124, ...
Connecting to packages-dev.wazuh.com (packages-dev.wazuh.com)|52.84.66.65|:443... connected.
HTTP request sent, awaiting response... 200 OK
Length: 705928884 (673M) [binary/octet-stream]
Saving to: ‘wazuh-indexer-4.6.0-wp.2420.x86_64.rpm’

100%[====================================================================>] 705,928,884 24.8MB/s   in 29s    

…